requirements.txt
langchain
openai
python-dotenv
streamlit
streamlit-chat
